update [ci skip]

+++ src/threepp/math/MathUtils.cpp
@@ -17,7 +17,7 @@ std::string math::generateUUID() {
     static std::uniform_real_distribution<double> dist(0.0, 1.0);
 
     const auto d0 = static_cast<size_t>(dist(e2) * 0xffffffff) | 0;
-    const auto d1 = static_cast<int>(dist(e2) * 0xffffffff) | 0;
+    const auto d1 = static_cast<size_t>(dist(e2) * 0xffffffff) | 0;
     const auto d2 = static_cast<size_t>(dist(e2) * 0xffffffff) | 0;
     const auto d3 = static_cast<size_t>(dist(e2) * 0xffffffff) | 0;
